- refactor for another project

+++ backend/db/index.js
@@ -49,10 +49,10 @@ function Storage() {
         connection.query('SELECT u.id, u.login FROM User u INNER JOIN Room_has_User ru ON u.id = ru.user_fk WHERE ru.room_fk =' + roomId, callback);
     };
 
-    API.deleteRoomById = function(roomId,callback) {
+    API.deleteRoomById = function(roomId,userId,callback) {
         connection.query('DELETE FROM Room WHERE id = ' + roomId);
         connection.query('DELETE FROM Product WHERE room_fk = ' + roomId);
-        connection.query('DELETE FROM Room_has_User WHERE room_fk = ' + roomId, callback);
+        connection.query('DELETE FROM Room_has_User WHERE room_fk = ' + roomId + ' AND user_fk = ' + userId, callback);
     };
 
     API.updateRoom = function (room, callback) {
